app/requirements.txt
Django
pillow
psycopg2-binary
djangorestframework
markdown
django-filter
numpy
pandas
matplotlib
